\name{psp21}
\alias{psp21}
\docType{data}
\title{
Czech 2021 parliamentary election results
}
\description{
A dataset containing vote totals of all 22 parties, movements, and coalitions which contested the 2021 parliamentary election in the Czech Republic.
}
\usage{data("psp21")}
\format{
  A data frame with 22 observations on the following 16 variables.
  \describe{
    \item{\code{Parties}}{Party names.}
    \item{\samp{Prague Capital Region}}{a numeric vector}
    \item{\samp{Central Bohemian Region}}{a numeric vector}
    \item{\samp{South Bohemian Region}}{a numeric vector}
    \item{\samp{Plzeň Region}}{a numeric vector}
    \item{\samp{Karlovy Vary Region}}{a numeric vector}
    \item{\samp{Ústí nad Labem Region}}{a numeric vector}
    \item{\samp{Liberec Region}}{a numeric vector}
    \item{\samp{Hradec Králové Region}}{a numeric vector}
    \item{\samp{Pardubice Region}}{a numeric vector}
    \item{\samp{Vysočina Region}}{a numeric vector}
    \item{\samp{South Moravian Region}}{a numeric vector}
    \item{\samp{Olomouc Region}}{a numeric vector}
    \item{\samp{Zlín Region}}{a numeric vector}
    \item{\samp{Moravian-Silesian Region}}{a numeric vector}
    \item{\code{Total}}{a numeric vector}
  }
}
\source{
Source: Czech Statistical Bureau, 2021
}
\examples{
data(psp21)
}
\keyword{datasets}
